lot of bug fix adding the tutorial

+++ src/utils/pagination.rs
@@ -1,7 +1,7 @@
 
 /// Convenient function to default pagination limit and index to default if none are provided.
 /// The provided values are clamped in the correct range.
-pub(crate) fn page_limit_and_index(page_limit: Option<usize>, page_index: Option<usize>) -> (usize, usize) {
+pub(crate) fn page_limit_and_index(page_limit: Option<u64>, page_index: Option<u64>) -> (u64, u64) {
     (
         match page_limit {
             Some(limit) => limit.min(20).max(1),
